codegen: Porting `rewrite_move` function into `rewrite` module.
